refactor: tweak form styling
