+++ src/index.scss
@@ -1,9 +1,11 @@
+@import "scss";
+
 :root {
     font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
     line-height: 1.5;
 
-    color: rgba(255, 255, 255, 0.87);
-    background-color: #242424;
+    color: $clr-secondary;
+    background-color: $clr-dark;
 }
 
 body {
