+++ src/modules/auth/components/forms/index.ts
@@ -1,0 +1,3 @@
+export * from "./login-form";
+export * from "./sign-up-form";
+export * from "./verify-mail-form";
