--- src/modules/auth/components/login-form/login-form.module.scss
@@ -1,7 +1,0 @@
-.form {
-    color: white;
-
-    label[for] {
-        color: white;
-    }
-}
