+++ src/modules/auth/components/password-input/password-input.module.scss
@@ -1,3 +1,5 @@
+@use "../../../../scss" as *;
+
 .toggleButton {
     border: 0;
     background: transparent;
@@ -6,6 +8,11 @@
     margin: 0;
 
     width: 1rem;
+    height: 1rem;
+
+    display: flex;
+    align-items: center;
+    justify-content: center;
 
     &__icon {
         svg {
@@ -13,6 +20,6 @@
             height: 1rem;
         }
 
-        color: #0003;
+        color: $clr-dark;
     }
 }
