--- src/modules/auth/components/sign-up-form/index.ts
@@ -1,1 +1,0 @@
-export * from "./sign-up-form";
