--- src/modules/auth/components/sign-up-form/sign-up-form.module.scss
@@ -1,7 +1,0 @@
-.form {
-    color: white;
-
-    label[for] {
-        color: white;
-    }
-}
