--- src/modules/auth/components/verify-mail-form/index.ts
@@ -1,1 +1,0 @@
-export * from "./verify-mail-form";
