--- src/modules/auth/components/verify-mail-form/verify-form.module.scss
@@ -1,7 +1,0 @@
-.form {
-    color: white;
-
-    label[for] {
-        color: white;
-    }
-}
