+++ src/pages/login/login.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from "@/modules/auth/components/login-form/login-form";
+import { LoginForm } from "@/modules/auth/components/forms";
 
 export const Login = () => {
     return <LoginForm />;
