+++ src/pages/verify-mail/verify-mail.tsx
@@ -1,4 +1,4 @@
-import { VerifyMailForm } from "@/modules/auth/components/verify-mail-form";
+import { VerifyMailForm } from "@/modules/auth/components/forms";
 
 export const VerifyMail = () => {
     return <VerifyMailForm />;
