+++ src/scss/variables/_colors.scss
@@ -1,4 +1,4 @@
 $clr-primary: rgb(198, 47, 49);
 $clr-secondary: rgb(255, 255, 255);
 $clr-secondary--darker: rgb(184, 184, 184);
-$clr-dark: rgb(53, 53, 53);
+$clr-dark: #2e2e2e;
